xrCore, level 3 warnings elimination, part 1

+++ src/xrCore/Animation/interp.cpp
@@ -268,7 +268,7 @@ float evalEnvelope(CEnvelope* env, float time)
         return env->keys[0]->value;
 
     // find the first and last keys
-    int sz = env->keys.size();
+    size_t sz = env->keys.size();
     skey = env->keys[0];
     ekey = env->keys[sz - 1];
     skey_n = env->keys[1];
